move sysapi_query to starsystem.py (kinda the best place for it,rlly)
